rename component because its used in add and remove
